Fix 6 integration bugs: booking vacancy, checkout SQL, timezone, concurrency
1. server.js: Add isVacantStatus() helper — legacy admin UI writes 'available'
   to JSONB blob; booking hold/submit/public-rooms all require 'vacant'. Now
   accepts 'available'/'empty'/'free' as valid bookable statuses everywhere
   (hold check, submit check, approve-and-assign, public rooms feed).

2. routes/tenant-ops.js: Fix checkout 42P08 error — CASE WHEN $2 IS NOT NULL
   on a numeric parameter is ambiguous to PG. Changed to $2::numeric IS NOT NULL.

3. db/pool.js: Set timezone='Asia/Bangkok' on every pool.connect() event so
   CURRENT_DATE/NOW() in scheduler SQL (late-fee, reminders, billing) matches
   JS Date which runs in ICT (UTC+7). Without this, SQL and JS disagreed by 7h.

4. services/features.js: Wrap features.save() in BEGIN/SELECT FOR UPDATE/COMMIT
   to prevent lost-update when two admins save feature flags simultaneously.

5. server.js: Wrap PUT /api/data/:key in SELECT FOR UPDATE transaction for
   critical blobs (rooms, config) to prevent last-write-wins on concurrent saves.

6. server.js: Block disabling citizenIdEncryption when encrypted rows exist —
   turning it off would leave existing AES-GCM ciphertext displayed as-is,
   corrupting the citizen ID display for every existing active tenant.

All 822 tests pass.

+++ routes/tenant-ops.js
@@ -2048,7 +2048,7 @@ module.exports = function buildTenantOpsRouter(ctx) {
               closed_reason = $3,
               closed_type = 'tenant_checkout',
               deposit_returned = $2,
-              deposit_returned_at = CASE WHEN $2 IS NOT NULL THEN NOW() ELSE NULL END,
+              deposit_returned_at = CASE WHEN $2::numeric IS NOT NULL THEN NOW() ELSE NULL END,
               deposit_return_reason = $3
              WHERE tenant_id=$1 AND status='active'
            RETURNING id, contract_no, room_id, start_date, monthly_rent, deposit, discount_pct`,
